move category option extensions to dhislogic
